+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ##ADDED/[0.0.14] - 12/09/23
 
+- New functions added to the ScrabbleGame class;
+
+---------------
+##ADDED/[0.0.14] - 12/09/23
+
 - New functions added to the Player class;
 - New functions added to the Dictionary class;